moved some fixtures to conftest
